Create SearchScreenContainer to contain logic for SearchScreen
