+++ App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {NavigationContainer} from '@react-navigation/native';
 import {createStackNavigator} from '@react-navigation/stack';
 import HomeScreen from './src/screens/HomeScreen';
-import SearchScreen from './src/screens/SearchScreen';
+import SearchScreen from './src/screens/SearchScreen/SearchScreen';
 
 const Stack = createStackNavigator();
