+++ src/screens/SearchScreen/SearchScreen.js
@@ -1,0 +1,10 @@
+import React, {useState} from 'react';
+import {View} from 'react-native';
+import SearchScreenContainer from './components/SearchScreenContainer';
+import {SearchSource} from 'jest';
+
+const SearchScreen = ({navigation}) => {
+  return <SearchScreenContainer />;
+};
+
+export default SearchScreen;
