+++ src/screens/SearchScreen/components/DictionaryEntriesContainer.js
