+++ src/screens/SearchScreen/components/DictionaryEntry.js
